requirements.txt
azure-common==1.1.23
azure-mgmt-hdinsight==1.2.0
azure-mgmt-resource==5.1.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
msrest==0.6.10
msrestazure==0.6.2
